Add simple JSX framework to each component
