--- components/Header.js
@@ -1,11 +1,0 @@
-import React from "react";
-
-function Header() {
-
-  
-  return (
-    
-  );
-}
-
-export default Header;
