--- components/PictureCard.js
@@ -1,11 +1,0 @@
-import React from "react";
-
-function PictureCard() {
-
-  
-  return (
-    
-  );
-}
-
-export default PictureCard;
