--- components/PreviousDates.js
@@ -1,11 +1,0 @@
-import React from "react";
-
-function PreviousDates() {
-
-  
-  return (
-    
-  );
-}
-
-export default PreviousDates;
