+++ src/components/Header.js
@@ -1,0 +1,14 @@
+import React from "react";
+
+function Header() {
+
+  
+  return (
+    <div>
+        <h1>NASA's Photo of the Day </h1>
+        <span>  ... date ... </span>
+    </div>
+  );
+}
+
+export default Header;
