+++ src/components/PictureCard.js
@@ -1,0 +1,18 @@
+import React from "react";
+
+function PictureCard() {
+
+  
+  return (
+    <div>
+      <h2>Photo title</h2>
+      <img src=""/>
+      <div>
+        <h3>Photo info</h3>
+        <p>info goes here...</p>
+      </div>
+    </div>
+  );
+}
+
+export default PictureCard;
