+++ src/components/PreviousDates.js
@@ -1,0 +1,17 @@
+import React from "react";
+import PreviousPictureCard from './PreviousPictureCard';
+
+function PreviousDates() {
+
+  
+  return (
+    <div>
+      <PreviousPictureCard />
+      <PreviousPictureCard />
+      <PreviousPictureCard />
+      <PreviousPictureCard />
+    </div>
+  );
+}
+
+export default PreviousDates;
