+++ src/components/PreviousPictureCard.js
@@ -4,7 +4,11 @@ function PreviousPictureCards() {
 
   
   return (
-    
+    <div>
+      <img src=""/>
+      <span>Photo title</span>
+      <p>...date...</p>
+    </div>
   );
 }
